QEE: narrow every integer op result to its declared width
The QuickExecutionEngine holds every integer width in one long buffer, so a
tensor's actual width lives only in RuntimeTensor.DType. Ops that compute at 64
bits — a left shift, a bitwise or/xor of already-widened operands — left bits
set above the declared width. Those bits are invisible to a later add (exact mod
2^w either way) but not to a right shift, which pulls them straight back down
into the result.

The reachable consequence is host constant folding, whose folded values re-enter
the graph as CONSTANT nodes: a partially-folded uint32 chain (Threefry's rotate
is a shift pair) silently computed values that disagreed with the same chain
left unfolded. Narrowing is applied once in QuickOp.RunCompute, the shared tail
every op already passes through, so no op has to remember it; it is idempotent
for the ops that already narrowed.

Also fixes an init-side dtype mismatch this exposed: FastInitKeyedDraws emitted
the keyed draw's drawBase as an int64 constant while the op definition declares
it uint64.

Full Purpose=Coverage suite: 710/710.

+++ src/Shorokoo/Core/Nodes/Processors/Fast/FastInitKeyedDraws.cs
@@ -22,7 +22,7 @@ namespace Shorokoo.Core.Nodes.Processors.Fast
     /// ceiling and parallelizes on GPU).
     ///
     /// <para>The draw node is rewritten in place: its key input is the parameter's folded
-    /// init key as a <c>[2]</c> constant, its drawBase is the draw's ordinal within the
+    /// init key as a uint64 scalar constant, its drawBase is the draw's ordinal within the
     /// initializer (a distinct sub-stream per draw; every shipping initializer has exactly
     /// one, so ordinal 0 in practice), and its shape input and declared distribution
     /// bounds carry over — the initializer's downstream scaling math is unchanged.
@@ -116,16 +116,16 @@ namespace Shorokoo.Core.Nodes.Processors.Fast
                 var shapeInput = node.Inputs[0]
                     ?? throw new InvalidOperationException("Random init node has null shape input.");
 
-                // The parameter's own stream key as a [2] constant, and a distinct sub-stream
+                // The parameter's own stream key as a scalar constant, and a distinct sub-stream
                 // (drawBase = ordinal) per draw within one initializer. The constant is emitted
                 // per draw so it always sits in the draw's own control-flow scope.
                 var keyKey = AppendConstant(new OnnxTensorData<uint64>(
                     new Shape(System.Array.Empty<long>()),
                     OnnxUtils.CreateTensorValue(new Shape(System.Array.Empty<long>()), (ulong[])[streamKey])), newNodes);
 
-                var drawBaseKey = AppendConstant(new OnnxTensorData<int64>(
+                var drawBaseKey = AppendConstant(new OnnxTensorData<uint64>(
                     new Shape(Array.Empty<long>()),
-                    OnnxUtils.CreateTensorValue(new Shape(Array.Empty<long>()), (long[])[randomOrdinal])), newNodes);
+                    OnnxUtils.CreateTensorValue(new Shape(Array.Empty<long>()), (ulong[])[(ulong)randomOrdinal])), newNodes);
 
                 if (isBits)
                 {
